Revert "Initial checkin of GA CTA button tracking"
This reverts commit 0eaf9d331b06c5fa40ef105016814aea5aac5ea6.
