+++ src/applications/static-pages/analytics/index.js
@@ -3,6 +3,7 @@ import * as Sentry from '@sentry/browser';
 import addJumplinkListeners from './addJumpLinkListeners';
 import addQaSectionListeners from './addQaSectionListeners';
 import addTeaserListeners from './addTeaserListeners';
+import addButtonLinkListeners from './addButtonLinkListeners';
 
 /**
  * Use pageListenersMap.set(<page path>, <array of functions>) to register
@@ -33,6 +34,7 @@ function attachAnalytics() {
 
     // Global listeners
     addTeaserListeners();
+    addButtonLinkListeners();
   } catch (error) {
     Sentry.withScope(scope => {
       scope.setExtra('error', error);
